Cleans up factories
