--- lang/src/org/partiql/lang/eval/physical/operators/AggregateOperator.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2022 Amazon.com, Inc. or its affiliates.  All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at:
- *
- *      http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
- * language governing permissions and limitations under the License.
- */
-
-package org.partiql.lang.eval.physical.operators
-
-import org.partiql.lang.domains.PartiqlPhysical
-import org.partiql.lang.eval.physical.SetVariableFunc
-
-public interface AggregateOperator {
-    public fun create(
-        source: RelationExpression,
-        keys: List<CompiledGroupKey>,
-        functions: List<CompiledAggregateFunction>
-    ): RelationExpression
-}
-
-public class CompiledGroupKey(
-    val setGroupKeyVal: SetVariableFunc,
-    val value: ValueExpression,
-    val variable: PartiqlPhysical.VarDecl
-)
-
-public class CompiledAggregateFunction(
-    val name: String,
-    val setAggregateVal: SetVariableFunc,
-    val value: ValueExpression,
-    val quantifier: PartiqlPhysical.SetQuantifier,
-)
